move arithmetic code out into separate files
